pyproject.toml
[project]
name = "learn-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ipython>=9.4.0",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.29",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.4",
    "langsmith>=0.4.13",
    "numpy>=2.3.2",
    "ollama>=0.5.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
